test: added successful search for a single customer credit option test. Added required Fixture classes.

+++ src/test/kotlin/dio/spring_kotlin_rest_tdd/factory/CustomerFixture.kt
@@ -3,7 +3,6 @@ package dio.spring_kotlin_rest_tdd.factory
 import dio.spring_kotlin_rest_tdd.model.Address
 import dio.spring_kotlin_rest_tdd.model.Customer
 import jakarta.validation.constraints.Email
-import org.hibernate.validator.constraints.br.CPF
 import kotlin.random.Random
 
 class CustomerFixture {
@@ -16,7 +15,7 @@ class CustomerFixture {
       cpf: String = "12345678935",
       email: String = "${firstName.lowercase()}@email.com",
       cep: String = "95000000",
-      income: Int = 100000,
+      income: Long = 100000,
       address: Address
     ): Customer {
       return Customer(id = id, firstName = firstName, lastName = lastName, cpf = cpf, email = email, cep = cep, address = address, income = income)
